Add support for CMake build on Windows
A CMake build file is now present along with a windows batch
file to speed up the build process. Documentation in both files
explains the use of environment variables that can/will need to
be configured by users of the build. This information should be
added to build documentation.

IMPORTANT NOTES:

* Only Tobii SDK 1.2.0.33 is supported. The latest version
of the SDK changes the folder structure and the names of the dlls.
* iTrace core can only be built in 64-bit release mode (for now).