Allow EyeTracker to gracefully stop
Update tracker to support a function to stop tracking
and force mouse and eyetracker classes to use this
interface. Adding support for tobii API function
"tobii_research_unsubscribe_from_gaze_data" turns off
listening on the callback function.

Tobii Tracker still needs porting to NOT read from dll
and use API. Additionally, threading issue still exists,
but this is NOT from the tobii api (which was causing the
crashing before).