Modify iTrace-Core to support CMake (#17)
This commit changes the way iTrace-Core is built.

1) The tobii header files have been removed from the project source
directory and are now an external dependency. A zip called
itrace-core_deps.zip will contain dependencies and is organized to
make windows build environments more consistent. The deps folder has
been added to the .gitignore to avoid the accidental adding of its
contents to the repository.

2) tobiipro.cpp still dynamically loads the tobii_research dll, but
supports proper loading when built via visual studio from GUI or CLI
using CMake.

3) Build.md outlines the changes to the build procedure and outlines
the steps and current limitations.

4) win_build.bat is an example for quick CLI builds using CMake.