Allow XML to support event time
Timestamps for xml data are now consolidated to gaze_data.hpp.
This does not include tracker time as that comes directly from
tracker library. Additional timestamps are as follows:

    * system_time = UNIX style timestamp (seconds since epoch)
    * event_time = Steady clock with nanosecond resolution
    * tracker_time = recorded from tracker hardware API

Event time is the unique identifier for each gaze response and
used to determine fixation durations.

Core has also been adjusted to transmit event_time and not
system time to the iTrace plugins. This field will allow for
the unification between core and plugin data.