Basic support for gazedata collection and buffering
Producer/Consumer approach to gaze data processing. This mimics the
inital approach from the C++ QT version. It is unclear at this point
if additional refactoring with Events/Delegates for this is beneficial
or necessary. Transmitting data to internal Core components
(Web/Standard Sockets, XML output, UI elements, etc.) is not yet
implemented.