Connect Session Setup Form to Manager
Data from form is now stored in session manager.
The button presses don't really animate in a  way that
indicates an action has occurred, so consider adding a
dialoge that indicates the data has been stored. Clear
has a more immediate effect, but perhaps a prompt to
see if the user actually wants to clear the data might be
handy.