Update window title bar text
Windows were displaying the title as the camelcase version
of the classname and look silly. They are now normal titles
and the word "Window" is removed due to it being redundant.
Need to test the windows that are only available when a tracker
is plugged in to fully verify the fix and generate a new installer.